refactor: split worker-related files into a sub-module
This allows a better organization of the code
